AHN-2781 Extracted `rdapConformance` handling into `TopMostEntityTrait` and applied it to relevant classes. Adjusted serialization logic and updated tests accordingly.
